Move SummaryState to a separate file
